RDKB-31421 : Warning error in CcspPandM
Reason for change: To resolve warnings in CcspPandM
Below listed warning types are fixed:
-Werror=unused-variable
-Werror=unused-but-set-variable
-Werror=format
-Werror=format-extra-args
-Werror=implicit-function-declaration
-Werror=pointer-sign
-Werror=missing-braces
-Werror=unused-function
-Werror=unused-value
incompatible pointer type
discards 'const' qualifier from pointer target type
missing whitespace after the macro name
assignment from pointer without a cast
Test Procedure: Basic functionalities of CcspPandM should be tested,regression test should pass.
Risks: High

(cherry picked from commit 3210254b764182e4f30bbe67d9bf28c38faad13c)

BWGRDK-1221: Unable to configure true static IP from MSO page

Reason for change: To solve CcspPandMSsp crash
Test Procedure: Verify that configuring true static IP from MSO page is successful and there are no CcspPandMSsp crashes.
Risks: Low

(cherry picked from commit 7f48a780ad1e6e7cf806ddecfeda8d9457186ef6)

RDKB-33080 : System and event logs are not generated in GUI.

Reason for change:
Looks like readdir return check not proper due to this syslog and
eventlog directories contents are missed to read. so needs to correct
it.

Test Procedure:
1. System log and Event log should display properly in GUI

Risks: Low

(cherry picked from commit 0948882dfac73b53cae27f30b88a9f2c07f4a021)

+++ source/TR-181/middle_layer_src/array_helpers.c
@@ -66,7 +66,7 @@ void* helper_convert_array( const void *buf, size_t len,
 
             /* The outermost wrapper is not present */
             mp_rv = msgpack_unpack_next( &msg, (const char*) buf, len, &offset );
-			printf("mp_rv is %d MSGPACK_UNPACK_SUCCESS %d offset %lu\n", mp_rv, MSGPACK_UNPACK_SUCCESS, offset);
+			printf("mp_rv is %d MSGPACK_UNPACK_SUCCESS %d offset %lu\n", mp_rv, MSGPACK_UNPACK_SUCCESS, (long unsigned int)offset);
 			msgpack_object obj = msg.data;
 			msgpack_object_print(stdout, obj);
